+++ database/migrations/2025_11_26_080358_add_pem_to_ratings_table.php
@@ -12,9 +12,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('ratings', function (Blueprint $table) {
-             $table->unsignedBigInteger('peminjaman_id')->nullable();
-
+            if (! Schema::hasColumn('ratings', 'peminjaman_id')) {
+                $table->unsignedBigInteger('peminjaman_id')->nullable();
                 $table->foreign('peminjaman_id')->references('id')->on('peminjamen')->onDelete('set null');
+            }
         });
     }
